+++ src/components/json/parser/_1/_false.h
@@ -34,14 +34,14 @@ namespace qor { namespace components { namespace parser { namespace json {
     {
         public: _false(Parser* parser) :
             Sequence(parser,
-                new_ref<Specific>(parser, 0x66),
+                new_ref<Specific>(parser, (byte)0x66),
                 new_ref<Sequence>(parser,
-                    new_ref<Specific>(parser, 0x61),
+                    new_ref<Specific>(parser, (byte)0x61),
                     new_ref<Sequence>(parser,
-                        new_ref<Specific>(parser, 0x6c),
+                        new_ref<Specific>(parser, (byte)0x6c),
                         new_ref<Sequence>(parser,
-                            new_ref<Specific>(parser, 0x73),
-                            new_ref<Specific>(parser, 0x65)
+                            new_ref<Specific>(parser, (byte)0x73),
+                            new_ref<Specific>(parser, (byte)0x65)
                         )
                     )
                 ),
