+++ src/components/json/parser/_1/begin_object.h
@@ -37,7 +37,7 @@ namespace qor { namespace components { namespace parser { namespace json {
             Sequence(parser,
                 new_ref<ws>(parser),
                 new_ref<Sequence>(parser,
-                    new_ref<Specific>(parser, 0x7B),
+                    new_ref<Specific>(parser, (byte)0x7B),
                     new_ref<ws>(parser)
                 ),
             static_cast<uint64_t>(jsonToken::begin_object)){ }
