+++ src/components/json/parser/_3/name_separator.h
@@ -37,7 +37,7 @@ namespace qor { namespace components { namespace parser { namespace json {
             Sequence(parser,
                 new_ref<ws>(parser),
                 new_ref<Sequence>(parser,
-                    new_ref<Specific>(parser, 0x3A),
+                    new_ref<Specific>(parser, (byte)0x3A),
                     new_ref<ws>(parser)
                 ),
             static_cast<uint64_t>(jsonToken::name_separator)){ }
